% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{applyFilter}
\alias{applyFilter}
\title{Filter genes and samples in coexList object for a given threshold}
\usage{
applyFilter(cl, propGenes = 1)
}
\arguments{
\item{cl}{An object of class coexList.}

\item{propGenes}{The proportion of genes to keep. Will rank and select genes with the highest variance.}
}
\description{
A function to identify samples and genes to be filtered for network construction.
}
\examples{
\dontrun{
cl <- coexList(d1)
dim(cl@exprs)
cl <- applyFilter(cl, propGenes=0.5)
dim(cl@exprs)
}
}
