% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-methods.R
\name{as.S3hclust,hclust-method}
\alias{as.S3hclust,hclust-method}
\title{Convert a S4 hclust to S3 hclust}
\usage{
\S4method{as.S3hclust}{hclust}(object)
}
\arguments{
\item{object}{A S4 hclust object}
}
\value{
An S3 hclust object
}
\description{
Convert a S4 hclust to S3 hclust
}
