% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{as.S3hclust}
\alias{as.S3hclust}
\title{Convert a S4 hclust to S3 hclust}
\usage{
as.S3hclust(object)
}
\arguments{
\item{object}{An S4 hclust object}
}
\value{
An S3 hclust object
}
\description{
Convert a S4 hclust to S3 hclust
}
