% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcModuleEigengene.R
\name{calcModuleEigengenes}
\alias{calcModuleEigengenes}
\title{Calculate an adjacency matrix}
\usage{
calcModuleEigengenes(cl, deepSplit = 2, ...)
}
\arguments{
\item{cl}{An object of class coexList.}

\item{deepSplit}{integer in the range 0 to 4.
Provides a rough control over sensitivity to cluster splitting.
The higher the value, the more and smaller clusters will be produced.
See ?dynamicTreeCut::cutreeHybrid for more details.}

\item{...}{Parameters for cutreeHybrid::dynamicTreeCut}
}
\value{
A coexList object.
}
\description{
Calculate an adjacency matrix
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- coexList(counts = edat)
cl <- normCounts(cl)
cl <- calcAdjacency(cl, softPower=6)
cl <- calcTOM(cl)
cl <- calcTree(cl)
cl <- calcModuleEigengenes(cl)
str(cl@moduleEigengenes)
}
