% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcModuleMembership.R
\name{calcModuleMembership}
\alias{calcModuleMembership}
\title{Calculate co-expression module membership}
\usage{
calcModuleMembership(cl, ...)
}
\arguments{
\item{cl}{An object of class CoexList.}

\item{...}{Parameters for WGCNA::signedKME}
}
\value{
A CoexList object. This function adds data to the 'moduleMembership'
slot of a CoexList object.
}
\description{
Calculate co-expression module membership
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- CoexList(counts = edat)
cl <- normCounts(cl)
cl <- calcAdjacency(cl, softPower=6)
cl <- calcTOM(cl)
cl <- calcTree(cl)
cl <- calcModuleEigengenes(cl)
cl <- calcModuleMembership(cl)
mm <- getModuleMembership(cl)
mm[1:3, 1:3]
}
\seealso{
\code{\link[WGCNA]{signedKME}}
}
