% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softPower.R
\name{calcSoftPower}
\alias{calcSoftPower}
\title{Analysis of scale free topology for soft-thresholding}
\usage{
calcSoftPower(cl, RsquaredCut = 0.85, ...)
}
\arguments{
\item{cl}{A coexList class object.}

\item{RsquaredCut}{Desired minimum scale free topology fitting index R^2.}

\item{...}{Arguments passed to \code{WGCNA:::pickSoftThreshold()}}
}
\value{
A coexList class object.
}
\description{
The aim of this function is to help choose an appropriate
soft-thresholding power for network construction. Use \code{plotSoftPower} to
inspect results
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rnorm(ngenes*nsamples,mean=5,sd=2),ngenes,nsamples)
rownames(edat) <- 1:ngenes
cl <- coexList(counts = edat)
cl <- calcSoftPower(cl)
cl@powerEstimate
head(cl@fitIndices)

}
\seealso{
WGCNA::pickSoftThreshold
}
