% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTree.R
\name{calcTree}
\alias{calcTree}
\title{Cluster topological overlap matrix using fastcluster::hclust}
\usage{
calcTree(cl, ...)
}
\arguments{
\item{cl}{An object of class coexList.}

\item{...}{Parameters for fastcluster::hclust()}
}
\value{
A coexList object.
}
\description{
A fast wrapper function for running hclust on a coexList object
on a coexList object
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- coexList(counts = edat)
cl <- normCounts(cl)
cl <- calcAdjacency(cl, softPower=6)
cl <- calcTOM(cl)
cl <- calcTree(cl)
# To view hclust object
cl@geneTree
}
