% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{checkFilter}
\alias{checkFilter}
\title{Check to see how many genes or samples will be filtered for a given threshold}
\usage{
checkFilter(cl, propGenes = 1, ...)
}
\arguments{
\item{cl}{An object of class coexList.}

\item{propGenes}{The proportion of genes to keep. Will rank and select genes
with the highest variance.}

\item{...}{Arguments passed to \code{WGCNA::goodSamplesGenes}}
}
\description{
A function to identify samples and genes to be filtered for
network construction.
}
\examples{
\dontrun{
cl <- coexList(d1)
checkFilter(cl)
}
}
\seealso{
goodSamplesGenes
}
