% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{coexList-class}
\alias{coexList-class}
\alias{.coexList}
\title{An S4 class to represent the data used in a weighted-gene co-expression network analysis.}
\description{
An S4 class to represent the data used in a weighted-gene co-expression network analysis.
}
\section{Slots}{

\describe{
\item{\code{normCounts}}{A matrix of normalised counts.}

\item{\code{phenoData}}{A data.frame with variable names (samples, libraries)
as rows, description tags (e.g., unit of measurement) as columns.}

\item{\code{isFiltered}}{Logical of length one. Indicates if data have been filtered on variance.}

\item{\code{powerEstimate}}{A numeric of length one. An estimate of an appropriate
soft-thresholding power calculated usingWGCNA::pickSoftThreshold()}

\item{\code{fitIndices}}{A data.frame containing the fit indices for scale free
topology calculated using WGCNA::pickSoftThreshold().}

\item{\code{softPower}}{A numeric of length one. The user-selected soft power
to be used for calculating the co-expression adjacency matrix.}

\item{\code{networkType}}{A character vector of length one. A network type.
Allowed values are (unique abbreviations of) "unsigned", "signed",
"signed hybrid", "distance".}

\item{\code{adjacencyMat}}{Matrix. A WGCNA adjacency matrix generated
with WGCNA::adjacency().}

\item{\code{dissTOM}}{Matrix. A topological overlap matrix generated
with WGCNA::TOMsimilarity().}

\item{\code{geneTree}}{A S4 class hclust object as defined by tidycoex::hclust().}
}}

