% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexList-methods.R
\name{CoexList}
\alias{CoexList}
\title{An S4 class object to represent the data used in a weighted-gene co-expression network analysis.}
\usage{
CoexList(counts, ...)
}
\arguments{
\item{counts}{A matrix of gene expression data.}

\item{...}{Parameters for SummarizedExperiment.}
}
\value{
A CoexList class object.
}
\description{
An S4 class object to represent the data used in a weighted-gene co-expression network analysis.
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rnorm(ngenes*nsamples,mean=5,sd=2),ngenes,nsamples)
rownames(edat) <- 1:ngenes
CoexList(counts = edat)

}
