% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\docType{class}
\name{hclust-class}
\alias{hclust-class}
\title{An S4 class object for storing the S3 output of hclust, FlashClust or fastcluster}
\description{
An S4 class object for storing the S3 output of hclust, FlashClust or fastcluster
}
\section{Slots}{

\describe{
\item{\code{merge}}{Matrix as represented in the hclust S3 object}

\item{\code{height}}{Numeric  as represented in the hclust S3 object}

\item{\code{order}}{Integer as represented in the hclust S3 object}

\item{\code{labels}}{NULL as represented in the hclust S3 object}

\item{\code{method}}{Character as represented in the hclust S3 object}

\item{\code{call}}{Call represented in the hclust S3 object}

\item{\code{dist.method}}{NULL as represented in the hclust S3 object}
}}

