% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normCounts}
\alias{normCounts}
\title{Normalise RNA-seq count data for co-expression analyses}
\usage{
normCounts(cl)
}
\arguments{
\item{cl}{An object of class coexList.}
}
\description{
Normalise RNA-seq count data for co-expression analyses
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- coexList(counts = edat)
cl <- normCounts(cl)
cl@normCounts[1:6, 1:6]

}
