% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModuleTree.R
\name{plotModuleTree}
\alias{plotModuleTree}
\title{Plot a dendrogram with modules}
\usage{
plotModuleTree(cl, ...)
}
\arguments{
\item{cl}{An object of class coexList.}

\item{...}{Parameters for fastcluster::hclust()}
}
\value{
A plot to the display device
}
\description{
Plot a dendrogram with modules
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- coexList(counts = edat)
cl <- normCounts(cl)
cl <- calcAdjacency(cl, softPower=6)
cl <- calcTOM(cl)
cl <- calcTree(cl)
plotModuleTree(cl)
}
