% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA of samples from normalised data}
\usage{
plotPCA(cl, colVar = "", x_pc = 1, y_pc = 2, scale = TRUE)
}
\arguments{
\item{cl}{An object of class coexList.}

\item{colVar}{Character or numeric that refers to a
column name of colData(cl)}

\item{x_pc}{numeric of length 1. The principal component to plot on the x-axis}

\item{y_pc}{numeric of length 1. The principal component to plot on the y-axis}

\item{scale}{logical. Should the data be scaled before
calculating principal components? See ?prcomp for more details.
A column name of colData(cl) for point colours.}
}
\value{
ggplot object
}
\description{
Plot PCA of samples from normalised data
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- coexList(counts = edat)
cl <- normCounts(cl)
plotPCA(cl)

}
