% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softPower.R
\name{plotSoftPower}
\alias{plotSoftPower}
\title{Plot scale free topology and mean connectivity for soft power thresholds.}
\usage{
plotSoftPower(cl, RsquaredCut = 0.85)
}
\arguments{
\item{cl}{An object of class CoexList.}

\item{RsquaredCut}{Desired minimum scale free topology fitting index R^2.}
}
\value{
A ggplot object.
}
\description{
Plot the results from \code{calcSoftPower}.
This function returns two plots: 1) Scale free topology model fit,
and 2) Mean connectivity against soft power thresholds.
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- CoexList(counts = edat)
cl <- normCounts(cl)
cl <- calcSoftPower(cl)
plotSoftPower(cl)

}
\seealso{
calcSoftPower
}
