% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoexList-methods.R
\name{show,CoexList-method}
\alias{show,CoexList-method}
\title{Show method for when calling a CoexList object}
\usage{
\S4method{show}{CoexList}(object)
}
\arguments{
\item{object}{A CoexList object}
}
\description{
Show method for when calling a CoexList object
}
