% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycoex-package.R
\docType{package}
\name{tidycoex-package}
\alias{tidycoex}
\alias{tidycoex-package}
\title{tidycoex: Tidy and fast implementation of WGCNA}
\description{
This package is designed to easily run a weighted gene co-expression network analysis.
}
\author{
\strong{Maintainer}: Sam Buckberry \email{sam.buckberry@telethonkids.org.au} (\href{https://orcid.org/0000-0003-2388-6046}{ORCID})

}
\keyword{internal}
