% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAdjacency.R
\name{calcAdjacency}
\alias{calcAdjacency}
\title{Calculate an adjacency matrix}
\usage{
calcAdjacency(
  cl,
  method = "wgcna",
  softPower = cl@powerEstimate,
  networkType = "signed",
  corFun = "pearson",
  ...
)
}
\arguments{
\item{cl}{An object of class CoexList.}

\item{method}{Character. The method to construct matrix.
Current options are 'wgcna', clr'. For 'wgcna', the function used is WGCNA::adjacency.
For 'clr', the function used is}

\item{softPower}{Integer. The soft thresholding power used to construct the
WGCNA network. This can be determined using the determineSoftPowerWGCNA function.}

\item{networkType}{character. The type of WGCNA network. Default is 'signed'.
See ?WGCNA::adjacency for options.}

\item{corFun}{character. The correlation method used.
Options are "pearson" or "spearman".}

\item{...}{Paramaters passed to adjacency function defined in method.}
}
\value{
A CoexList object.
}
\description{
A wrapper function for constructing a co-expression matrix
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- CoexList(counts = edat)
cl <- normCounts(cl)
cl <- calcAdjacency(cl, softPower=6)
# To view adjacency matrix
cl@adjacencyMat[1:6, 1:6]
}
