% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAdjacency.R
\name{calcAdjacency}
\alias{calcAdjacency}
\title{Calculate an adjacency matrix}
\usage{
calcAdjacency(
  cl,
  method = "wgcna",
  softPower = cl@powerEstimate,
  TOM = TRUE,
  networkType = "signed",
  corFun = "pearson"
)
}
\arguments{
\item{cl}{An object of class coexList.}

\item{method}{Character. The method to construct matrix.
Current options are 'wgcna', clr'.}

\item{softPower}{Integer. The soft thresholding power used to construct the
WGCNA network. This can be determined using the determineSoftPowerWGCNA function.}

\item{TOM}{Logical. Calculate Topological overlap matrix for clustering?
Default = TRUE}

\item{networkType}{A character vector of length one. A network type.
Allowed values are (unique abbreviations of) "unsigned", "signed",
"signed hybrid", "distance"}

\item{corFun}{Character. The correlation function to be used.
Can be one of pearson (default), or spearman.}
}
\value{
A coexList object.
}
\description{
A wrapper function for constructing a co-expression matrix
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- coexList(counts = edat)
cl <- normCounts(cl)
cl <- calcAdjacency(cl, softPower=6)
# To view adjacency matrix
cl@adjacencyMat[1:6, 1:6]
# To view topological overlap matrix
cl@dissTOM[1:6, 1:6]
}
