% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTOM.R
\name{calcTOM}
\alias{calcTOM}
\title{Calculate an adjacency matrix}
\usage{
calcTOM(cl, ...)
}
\arguments{
\item{cl}{An object of class CoexList.}

\item{...}{Parameters for WGCNA::TOMsimilarity()}
}
\value{
A CoexList object.
}
\description{
A wrapper function for running the WGCNA::TOMsimilarity function
on a CoexList object
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- CoexList(counts = edat)
cl <- normCounts(cl)
cl <- calcAdjacency(cl, softPower=6)
cl <- calcTOM(cl)
# To view TOM
cl@dissTOM[1:6, 1:6]
}
