% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleMembership.R
\name{getModuleMembership}
\alias{getModuleMembership}
\title{Get co-expression module membership data from coexList object}
\usage{
getModuleMembership(cl)
}
\arguments{
\item{cl}{An object of class coexList.}

\item{...}{Parameters for WGCNA::signedKME}
}
\value{
A coexList object. This function adds data to the 'moduleMembership'
slot of a coexList object.
}
\description{
Get co-expression module membership data from coexList object
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- coexList(counts = edat)
cl <- normCounts(cl)
cl <- calcAdjacency(cl, softPower=6)
cl <- calcTOM(cl)
cl <- calcTree(cl)
cl <- calcModuleEigengenes(cl)
cl <- calcModuleMembership(cl)
mm <- getModuleMembership(cl)
mm[1:3, 1:3]

}
\seealso{
\code{\link[WGCNA]{signedKME}}
}
