% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normCounts}
\alias{normCounts}
\title{Normalise RNA-seq count data for co-expression analyses}
\usage{
normCounts(cl, normMethod = "CPM", scaleMethod = "log2", geneLength = NA)
}
\arguments{
\item{cl}{An object of class CoexList.}

\item{normMethod}{character. The within-sample normalisation method.
One of "CTF", "CPM", "TPM"}

\item{scaleMethod}{character. One of "log2", "asinh",}

\item{geneLength}{A numeric vector of length nrow(cl)
that is the length of genes for length based normalisation such as TPM.}
}
\description{
Normalise RNA-seq count data for co-expression analyses
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- CoexList(counts = edat)
cl <- normCounts(cl)
cl@normCounts[1:6, 1:6]
cl@normMethod
cl@scaleMethod

}
