% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModuleTree.R
\name{plotModuleTree}
\alias{plotModuleTree}
\title{Plot a dendrogram with modules}
\usage{
plotModuleTree(cl, deepSplits = c(0:4), cores = 1, ...)
}
\arguments{
\item{cl}{An object of class CoexList.}

\item{deepSplits}{integer in the range 0 to 4.
Provides a rough control over sensitivity to cluster splitting.
The higher the value, the more and smaller clusters will be produced.
See dynamicTreeCut::cutreeHybrid() for more information.}

\item{cores}{integer 1-5. How many cores to use?
Parameter passed to parallel::mclapply(mc.cores=cores)}

\item{...}{Parameters for dynamicTreeCut::cutreeHybrid()}
}
\value{
A plot to the display device
}
\description{
Plot a dendrogram with modules
}
\examples{
ngenes <- 1000
nsamples <- 16
edat <- matrix(rpois(ngenes*nsamples, lambda=5), nrow=ngenes)
rownames(edat) <- paste0("gene_", 1:ngenes)
cl <- CoexList(counts = edat)
cl <- normCounts(cl)
cl <- calcAdjacency(cl, softPower=6)
cl <- calcTOM(cl)
cl <- calcTree(cl)
plotModuleTree(cl)
}
