% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexList-methods.R
\name{show,coexList-method}
\alias{show,coexList-method}
\title{Show method for when calling a coexList object}
\usage{
\S4method{show}{coexList}(object)
}
\arguments{
\item{object}{A coexList object}
}
\description{
Show method for when calling a coexList object
}
